\name{cyt}
\docType{data}
\alias{cyt}

\title{Concha y Toro vineyards}
\description{
  Monthly returns of Concha y Toro vineyard collected over the period among the
  months of March 1990 to April 1999.
}
\usage{data(cyt)}
\format{
  A data frame with 110 observations on the following 3 variables.
  \describe{
    \item{Date}{ the month in which the observations were collected.}
    \item{IPSA}{ monthly returns for the Chilean market index.}
    \item{CyT}{ monthly returns, adjusted by equity variations, of Concha y Toro Vineyards.}
  }
}
\source{
  Osorio, F., and Galea, M. (2006).
  Detection of a change-point in student-t linear regression models.
  \emph{Statistical Papers} \bold{47}, 31-48.
}
\keyword{datasets}
