\name{dental}
\docType{data}
\alias{dental}

\title{An orthodontic measurement on several children over time}
\description{
  The \code{dental} data set introduced by Potthoff and Roy (1964), where dental
  measurements were made on 11 girls and 16 boys at ages 8, 10, 12 and 14.
  The response variable was the distance, in millimeters, from the center of
  pituitary to the pterygomaxillary fissure.
}
\usage{data(dental)}
\format{data frame with 108 rows and 4 columns.}
\source{
  Potthoff, R.F., and Roy, S.N. (1964). 
  A generalized multivariate analysis model useful especially for growth curve problems.
  \emph{Biometrika} \bold{51}, 313-326.
}
\keyword{datasets}
