\name{dialyzer}
\docType{data}
\alias{dialyzer}

\title{Ultrafiltration measurements for three centres}
\description{
  Data from a study to evaluate the \emph{in vivo} ultrafiltration characteristics 
  of a group of hollow fiber dialyzers. The dializers were evaluated in three centres
  with each \code{centre} using a different type of dialysate delivery system. The
  data set was introduced by Vonesh and Carter (1987) consists of four ultrafiltration
  rate measurements (\code{y1}, \code{y2}, \code{y3} and \code{y4}).
}
\usage{data(dialyzer)}
\format{A data frame with 40 rows and 5 columns.}
\source{
  Munoz-Pichardo, J.M., Moreno-Rebollo, J.L., Enguix-Gonzalez, A., and Pascual-Acosta, A. (2008).
  Influence measures on profile analysis with elliptical data through Frechet's metric.
  \emph{Metrika} \bold{68}, 111-127.

  Vonesh, E.F., and Carter, R.L. (1987).
  Efficient inference for random-coefficient growth curve models with unbalanced data.
  \emph{Biometrics} \bold{43}, 617-628.
}
\keyword{datasets}
