\name{housePrices}
\docType{data}
\alias{housePrices}

\title{Hedonic housing-price equation}
\description{
  Harrison and Rubinfeld (1978) reported a study of house prices in the Boston
  Metropolitan area. Their principal goal was to assess the association of the
  air quality in the neighbourhood with the house price after taking account of
  other influences.
}
\usage{data(housePrices)}
\format{data frame with 506 rows and 15 columns.}
\source{
  Harrison, D., Rubinfeld, D.L. (1978).
  Hedonic prices and the demand for clean air.
  \emph{Journal of Environmental Economics and Management} \bold{5}, 81-102.
}
\keyword{datasets}
