\name{thermocouples}
\docType{data}
\alias{thermocouples}

\title{Precision and accuracy of thermocouples}
\description{
  Christensen and Blackwood (1993) reported 64 sets of simultaneous measurements
  for temperature obtained by five thermocouples previously used, with the aim of
  examining their precision and exactitude after one or more thermocouples had
  sustain certain damage.
}
\usage{data(thermocouples)}
\format{data frame with 64 rows and 5 columns.}
\source{
  Christensen, R., Blackwood, L.G. (1993).
  Test for precision and accuracy of multiple measuring devices.
  \emph{Technometrics} \bold{35}, 411-420.
}
\keyword{datasets}
