\name{cyt}
\docType{data}
\alias{cyt}

\title{Concha y Toro Vineyards}
\description{
  Monthly returns of Concha y Toro Vineyard collected over the period among the
  months of March 1990 to April 1999.
}
\usage{data(cyt)}
\format{
  A data frame with 110 observations on the following 3 variables.
  \describe{
    \item{Date}{the month the data was collected}
    \item{IPSA}{return for the Chilean market}
    \item{CyT}{monthly returns, adjusted by equity variations, of Concha y Toro Vineyards}
  }
}
\source{
  Osorio, F. and Galea, M. (2005).
  Detection of a change-point in student-t linear regression models.
  \emph{Statistical Papers} \bold{45}, 31-48.
}
\keyword{datasets}
